+++ changelog.md
@@ -1,2 +1,3 @@
-Volucraft 0.0.5+26.1.2
-- Slots are now selectable in RRV's RecipeViewScreen
+Volucraft 0.0.6+26.1.2
+- Limit RRV's amalgamation recipe displays to one per Menu/Screen
+- Create actual crafting recipe for Amalgamation Table